Mengubah isi point jadi int, membuat adt list dinamis yang berisi label dan point

+++ src/adt/point.h
@@ -9,8 +9,8 @@
 
 typedef struct
 {
-   float X; /* absis   */
-   float Y; /* ordinat */
+   int X; /* absis   */
+   int Y; /* ordinat */
 } POINT;
 
 /* *** Notasi Akses: Selektor POINT *** */
